requirements.txt
Django==5.1.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
sqlparse==0.5.2
requests==2.32.3
python-dotenv==1.0.1
kokoro==0.7.14
numpy==1.26.4
scipy==1.15.1
joblib==1.4.2
tqdm==4.67.1
spacy==3.8.4
thinc==8.3.4
transformers==4.48.1
tokenizers==0.21.0
huggingface-hub==0.28.0
torch==2.6.0
soundfile
Pillow
iPython
huggingface-hub==0.28.0 
misaki==0.7.14 
sqlparse==0.5.2 
typer==0.15.1
web3
eth_account
mnemonic
py-solc-x
